peridot: parts: Introduce Display Satutaion Service and TileUtils
* Adjust Saturation Value Handling for Device Color Mode Compatibility

This update refines the saturation adjustment mechanism. When the seek bar progress is set, the method executes
the command "service call SurfaceFlinger 1022 f [saturation]" via Runtime.getRuntime().exec() to apply the saturation changes.
However, it was observed that when the device's color mode is set to "saturated," a saturation value of 1.0 does not trigger any visual changes.
To address this, the code now sets a value of 1.001 for a seek bar value of 100, ensuring the desired effect is consistently applied.

* TileUtils Implementation: Implemented TileUtils to handle the addition of Quick Settings tiles, specifically for the saturation feature.
This is done with reference to the guidelines provided by Google
https://developer.android.com/develop/ui/views/quicksettings-tiles#prompt-user
this ensuring a seamless integration and user experience when creating new tiles.

Signed-off-by: kenway214 <kenway214@outlook.com>

peridot: parts: [FIX] Define missing continuousUpdates attribute

- Added the `continuousUpdates` attribute to `attrs.xml` to resolve the XML parsing error in `saturation.xml`.
- This ensures the correct functioning of the saturation feature.

Signed-off-by: kenway214 <kenway214@outlook.com>

peridot: parts: Add XLIFF namespace to prevent XML parsing errors

- Added `xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2"` to the resources tag to resolve parsing errors related to the use of <xliff:g> tags in string resources.
- This ensures that the custom seekbar strings work correctly with the necessary placeholders.

* while we are there also adding `xmlns:tools` namespace declaration to AndroidManifest

peridot: parts: Rewrite in kotlin

Signed-off-by: kenway214 <kenway214@outlook.com>
Co-Authored-By: AnierinB <anierin@evolution-x.org>
Signed-off-by: Drenzzz <realdrenzzz@gmail.com>
Signed-off-by: Shikhar <Krtonia@gmail.com>