peridot: wifi: Use default value for assoc failure timeout
Using the value from xiaomi changes is out of bounds since [1]
it reduces the value limit, so if using the value from xiaomi
it will fallback and use the default value. If we use the max
value it causes internet connection problems when using wifi,
so let it use the default value.

[1]: https://git.codelinaro.org/clo/la/platform/vendor/qcom-opensource/wlan/qcacld-3.0/-/commit/b64278b0a56668077f24f4334dcf125bbadf9642

Change-Id: I7f6ef4d96ce40d123526ae1b951b544f52d28ff9
Signed-off-by: Fiqri Ardyansyah <fiqri191002@gmail.com>
Signed-off-by: Shikhar <Krtonia@gmail.com>